+++ src/types/next-auth.d.ts
@@ -4,6 +4,11 @@ declare module 'next-auth' {
 	interface Session {
 		sessionToken?: string | null;
 		currentOrgId?: string | null;
+		orgId?: string | null;
+		role?: string | null;
+		currentCompanyId?: string | null;
+		companyId?: string | null;
+		companyRole?: string | null;
 		user: {
 			id: string;
 			name?: string | null;